+++ ChangeLog.md
@@ -3,6 +3,9 @@
 ## 11/10/2021 2.10.3
 
 - Fix 'generate-examples' command failed to generate correct format for ‘uri’ format parameter
+- LiveValidatorLoader - make response headers as optional for swagger 2.0
+- LiveValidator - ignore LRO header check if using 201/200+ provisioningState
+- ModelValidator/SemanticValidator - support suppression on error code
 - Add new rules of 'INVALID_DISCRIMINATOR_VALUE' and 'DISCRIMINATOR_PROPERTY_NOT_FOUND'
 
 ## 11/04/2021 2.10.2